chore: MAT-177 경기 기록 페이지 컴포넌트 역할 분리로 코드 길이 개선 (347 -> 164)
