+++ src/routes/matches/$matchId/record/-components/MatchRecordLayout/MatchRecordLayout.tsx
@@ -6,8 +6,16 @@ import {
   teamAwayColor,
   teamHomeColor,
 } from '@/components/MatchLogList/colors.css';
+import { commonPaper } from '@/styles/paper.css';
 
 import * as styles from './MatchRecordLayout.css';
+
+// TO DO: 예시용 스타일로, 추후 제거 필요
+const s = (height: number | string) => ({
+  height,
+  backgroundColor: '#fff',
+  borderRadius: 10,
+});
 
 interface MatchRecordLayoutProps {
   team1: ReactNode;
@@ -42,17 +50,42 @@ export const MatchRecordLayout = ({
         [teamAwayColor]: team2Color,
       })}
     >
-      <div className={styles.teamContainer}>{team1}</div>
-      <div className={styles.centerContainer}>
-        {teamStats}
-        {selectedPlayer}
+      <div className={styles.teamContainer}>
+        <div
+          className={commonPaper}
+          style={{
+            ...s('auto'),
+            display: 'flex',
+            flexDirection: 'column',
+            gap: 8,
+          }}
+        >
+          {team1}
+        </div>
       </div>
-      <div className={styles.teamContainer}>{team2}</div>
+      <div className={styles.centerContainer}>
+        <div style={s(560)}>{teamStats}</div>
+        <div style={s(302)}>{selectedPlayer}</div>
+      </div>
+
+      <div className={styles.teamContainer}>
+        <div
+          className={commonPaper}
+          style={{
+            ...s('auto'),
+            display: 'flex',
+            flexDirection: 'column',
+            gap: 8,
+          }}
+        >
+          {team2}
+        </div>
+      </div>
       <div className={styles.infoContainer}>
-        {timer}
-        {info}
-        {log}
-        {memo}
+        <div style={s(116)}>{timer}</div>
+        <div style={s(284)}>{info}</div>
+        <div style={s(228)}>{log}</div>
+        <div style={s(204)}>{memo}</div>
       </div>
     </div>
   );
