+++ src/routes/matches/$matchId/record/-hooks/index.ts
@@ -1,0 +1,1 @@
+export * from './useMatchRecordWebSocket';
