+++ src/routes/matches/$matchId/record/-utils/index.ts
@@ -1,0 +1,1 @@
+export * from './dividePlayers';
