+++ src/styles/atomic.css.ts
@@ -1,0 +1,6 @@
+import { style } from '@vanilla-extract/css';
+
+export const flexContainer = style({
+  display: 'flex',
+  flexDirection: 'column',
+});
